Update Summary of C Language and Fahreheit Celsius Program

+++ fahrenheit_celsius.c
@@ -2,18 +2,18 @@
 /* print Fahrenheit-Celsius table
    for fahr = 0. 20, ..., 300 */
 main(){
-  int fahr, celsius;
-  int lower, upper, step;
+  float fahr, celsius;
+  float lower, upper, step;
 
-  lower = 0;
-  upper = 300;
-  step = 20;
+  lower = 0.0f;
+  upper = 300.0f;
+  step = 20.0f;
 
   fahr = lower;
   while (fahr<=upper){
 
-    celsius = 5*(fahr - 32)/9;
-    printf("%d\t%d\n", fahr,celsius );
+    celsius = 5.0*(fahr - 32)/9.0;
+    printf("%3.0f\t%5.2f\n", fahr,celsius );
 
     fahr += step;
   }
